Moving weights over, adjust configs
