+++ libs/types/src/tranches.rs
@@ -1,0 +1,43 @@
+use codec::{Decode, Encode};
+use frame_support::{traits::Get, BoundedVec, RuntimeDebug};
+use scale_info::TypeInfo;
+use sp_runtime::Perquintill;
+
+#[derive(Copy, Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+pub enum TrancheType<Rate> {
+	Residual,
+	NonResidual {
+		interest_rate_per_sec: Rate,
+		min_risk_buffer: Perquintill,
+	},
+}
+
+#[derive(Debug, Encode, PartialEq, Eq, Decode, Clone, TypeInfo)]
+pub struct TrancheMetadata<MaxTokenNameLength, MaxTokenSymbolLength>
+where
+	MaxTokenNameLength: Get<u32>,
+	MaxTokenSymbolLength: Get<u32>,
+{
+	pub token_name: BoundedVec<u8, MaxTokenNameLength>,
+	pub token_symbol: BoundedVec<u8, MaxTokenSymbolLength>,
+}
+
+/// Type that indicates the seniority of a tranche
+pub type Seniority = u32;
+
+#[derive(Debug, Encode, PartialEq, Eq, Decode, Clone, TypeInfo)]
+pub struct TrancheInput<Rate, MaxTokenNameLength, MaxTokenSymbolLength>
+where
+	MaxTokenNameLength: Get<u32>,
+	MaxTokenSymbolLength: Get<u32>,
+{
+	pub tranche_type: TrancheType<Rate>,
+	pub seniority: Option<Seniority>,
+	pub metadata: TrancheMetadata<MaxTokenNameLength, MaxTokenSymbolLength>,
+}
+
+#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+pub struct TrancheUpdate<Rate> {
+	pub tranche_type: TrancheType<Rate>,
+	pub seniority: Option<Seniority>,
+}
