+++ pallets/pools-registry/Cargo.toml
@@ -21,8 +21,10 @@ frame-system = { git = "https://github.com/paritytech/substrate", default-featur
 
 sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.26" }
 sp-std = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.26" }
+sp-arithmetic = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.26" }
 xcm = { git = "https://github.com/paritytech/polkadot", default-features = false , branch = "release-v0.9.26" }
 polkadot-parachain = { git = "https://github.com/paritytech/polkadot", default-features = false , branch = "release-v0.9.26" }
+orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.26" }
 orml-asset-registry = {  git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.26" }
 
 cfg-traits = { path = "../../libs/traits", default-features = false }
@@ -51,4 +53,5 @@ std = [
     'cfg-primitives/std',
     'polkadot-parachain/std',
     'xcm/std',
+    'orml-traits/std',
 ]
