+++ pallets/pools/src/weights.rs
@@ -30,8 +30,6 @@ use sp_std::marker::PhantomData;
 
 /// Weight functions needed for pallet_pools.
 pub trait WeightInfo {
-	fn create(n: u32) -> Weight;
-	fn update_no_execution(n: u32) -> Weight;
 	fn update_and_execute(n: u32) -> Weight;
 	fn execute_scheduled_update(n: u32) -> Weight;
 	fn set_max_reserve() -> Weight;
@@ -45,20 +43,6 @@ pub trait WeightInfo {
 /// Weights for pallet_pools using the Substrate node and recommended hardware.
 pub struct SubstrateWeight<T>(PhantomData<T>);
 impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
-	fn create(n: u32) -> Weight {
-		(Weight::from_ref_time(74_485_000)) // Standard Error: 41_000
-			.saturating_add((Weight::from_ref_time(299_000)).saturating_mul(n as u64))
-			.saturating_add(T::DbWeight::get().reads(6_u64))
-			.saturating_add(T::DbWeight::get().writes(6_u64))
-	}
-
-	fn update_no_execution(n: u32) -> Weight {
-		(Weight::from_ref_time(28_660_000)) // Standard Error: 17_000
-			.saturating_add((Weight::from_ref_time(285_000)).saturating_mul(n as u64))
-			.saturating_add(T::DbWeight::get().reads(4_u64))
-			.saturating_add(T::DbWeight::get().writes(1_u64))
-	}
-
 	fn update_and_execute(n: u32) -> Weight {
 		(Weight::from_ref_time(47_893_000)) // Standard Error: 40_000
 			.saturating_add((Weight::from_ref_time(716_000)).saturating_mul(n as u64))
@@ -123,20 +107,6 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 
 // For backwards compatibility and tests
 impl WeightInfo for () {
-	fn create(n: u32) -> Weight {
-		(Weight::from_ref_time(74_485_000)) // Standard Error: 41_000
-			.saturating_add((Weight::from_ref_time(299_000)).saturating_mul(n as u64))
-			.saturating_add(RocksDbWeight::get().reads(6_u64))
-			.saturating_add(RocksDbWeight::get().writes(6_u64))
-	}
-
-	fn update_no_execution(n: u32) -> Weight {
-		(Weight::from_ref_time(28_660_000)) // Standard Error: 17_000
-			.saturating_add((Weight::from_ref_time(285_000)).saturating_mul(n as u64))
-			.saturating_add(RocksDbWeight::get().reads(4_u64))
-			.saturating_add(RocksDbWeight::get().writes(1_u64))
-	}
-
 	fn update_and_execute(n: u32) -> Weight {
 		(Weight::from_ref_time(47_893_000)) // Standard Error: 40_000
 			.saturating_add((Weight::from_ref_time(716_000)).saturating_mul(n as u64))
